Refractored code to provide more clear code in most of the controllers

+++ src/main/java/com/server/socialBees/dto/WorkDTO.java
@@ -1,5 +1,6 @@
 package com.server.socialBees.dto;
 
+import com.server.socialBees.entity.Tag;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,7 +15,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class WorkDTO {
-    private Integer userId;
+    private Long userId;
     private String title;
     private String content;
     private LocalDate date;
