+++ src/main/java/com/server/socialBees/repository/CommentRepository.java
@@ -3,6 +3,6 @@ package com.server.socialBees.repository;
 import com.server.socialBees.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    Comment findCommentById(Integer commentId);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    Comment findCommentById(Long commentId);
 }
