+++ src/main/java/com/server/socialBees/repository/WorkRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Set;
 
-public interface WorkRepository extends JpaRepository<Work, Integer> {
-    Work findWorkById(Integer workId);
+public interface WorkRepository extends JpaRepository<Work, Long> {
+    Work findWorkById(Long workId);
     List<Work> findByTagsIn(Set<Tag> tags);
 }
