+++ src/main/java/com/server/socialBees/service/FileService.java
@@ -9,7 +9,7 @@ import java.util.stream.Stream;
 
 public interface FileService {
     List<FileDB> store(List<MultipartFile> files) throws IOException;
-    List<FileDB> updateFiles(List<MultipartFile> addedFiles, Integer workId) throws IOException;
+    List<FileDB> updateFiles(List<MultipartFile> addedFiles, Long workId) throws IOException;
     FileDB getFile(String id);
     Stream<FileDB> getAllFiles();
 }
