+++ src/main/java/com/server/socialBees/service/UserService.java
@@ -9,12 +9,12 @@ import java.util.Set;
 
 public interface UserService {
     User createUser(User user);
-    User getUserById(Integer userId);
+    User getUserById(Long userId);
     User getUserByEmail(String email);
     User updateUser(User newUser);
-    void followTag(Integer userId, Integer tagId);
-    void unfollowTag(Integer userId, Integer tagId);
-    List<Work> getWorksForUser(Integer userId);
-    Set<Tag> getFollowedTags(Integer userId);
-    User deleteUserById(Integer userId);
+    void followTag(Long userId, Long tagId);
+    void unfollowTag(Long userId, Long tagId);
+    List<Work> getWorksForUser(Long userId);
+    Set<Tag> getFollowedTags(Long userId);
+    User deleteUserById(Long userId);
 }
